Updated OrganizePhotos.py Code
Added a make directory function

Added two loops. One loop to create a list of place names, which are added to a list. This list is passed to the create directory function. 

The second loop is using the originals list (list of all photos), and then adding each photo in the list to the new directory accordingly.